Melhorias percebidas durante implementacao da classe em um projeto real.

+++ src/Elements/ICMSIPI/D001.php
@@ -23,7 +23,7 @@ class D001 extends Element implements ElementInterface
 
     protected $parameters = [
         'IND_MOV' => [
-            'type'     => 'integer',
+            'type'     => 'numeric',
             'regex'    => '^[0-1]{1}$',
             'required' => true,
             'info'     => 'Indicador de movimento: '
